+++ packages/components/pie-cookie-banner/test/accessibility/pie-cookie-banner.spec.ts
@@ -1,4 +1,3 @@
-
 import { test, expect } from '@justeattakeaway/pie-webc-testing/src/playwright/playwright-fixtures.ts';
 import { CookieBannerComponent } from '../helpers/page-object/pie-cookie-banner.page.ts';