+++ apps/examples/wc-next13/CHANGELOG.md
@@ -1,5 +1,12 @@
 # wc-next13
 
+## 0.4.24
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.5.38
+
 ## 0.4.23
 
 ### Patch Changes