+++ apps/examples/wc-nuxt2/CHANGELOG.md
@@ -1,5 +1,12 @@
 # wc-nuxt2
 
+## 0.7.25
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.5.38
+
 ## 0.7.24
 
 ### Patch Changes