+++ packages/components/pie-tag/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @justeattakeaway/pie-tag
 
+## 0.10.1
+
+### Patch Changes
+
+- Updated dependencies [[`b14535438`](https://github.com/justeattakeaway/pie/commit/b14535438da533aeb9a5520c2a239203dafb0c9a)]:
+  - @justeattakeaway/pie-webc-core@0.24.1
+
 ## 0.10.0
 
 ### Minor Changes