+++ packages/components/pie-toast/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @justeattakeaway/pie-toast
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [[`b14535438`](https://github.com/justeattakeaway/pie/commit/b14535438da533aeb9a5520c2a239203dafb0c9a)]:
+  - @justeattakeaway/pie-webc-core@0.24.1
+  - @justeattakeaway/pie-button@0.49.2
+  - @justeattakeaway/pie-icon-button@0.28.13
+  - @justeattakeaway/pie-icons-webc@0.25.2
+
 ## 0.3.3
 
 ### Patch Changes